requirements.txt
chess~=1.11
PyYAML~=6.0
requests~=2.32
backoff~=2.2
rich~=14.1
torch>=2.0.0
numpy~=1.26
groq
python-dotenv
fastapi
uvicorn
pydantic
